local cache instead of linking byte file into library binary
