+++ .env
@@ -1,0 +1,3 @@
+#! List of environment variables
+
+AZTEC_SRS_PATH="data/aztec20/kzg10-aztec20-srs-1024.bin"
