+++ src/bin/save_aztec_srs.rs
@@ -1,5 +1,17 @@
 //! Load Aztec SRS and save to files
 
 fn main() {
-    ark_srs::load::kzg10::bn254::aztec::store_aztec_srs(None).unwrap();
+    let degrees = [1024];
+
+    for degree in degrees {
+        print!("Parsing SRS for degree {degree} ...");
+        let mut srs = ark_srs::kzg10::aztec20::setup_from_raw(degree).unwrap();
+        print!(" done.\n");
+        srs.powers_of_g.truncate(degree + 1);
+
+        let dest = ark_srs::load::kzg10::bn254::aztec::default_path(degree).unwrap();
+        print!("Saving to {} ...", dest.display());
+        ark_srs::load::store_data(srs, dest).unwrap();
+        print!(" done.\n");
+    }
 }
