+++ src/constants.rs
@@ -1,0 +1,5 @@
+//! Configurations and constants, centralized for single responsibility
+
+/// data related to Aztec's ignition ceremony (including original transcripts
+/// from the ceremony and the arkworks serialized data blobs)
+pub(crate) const AZTEC20_DIR: &str = "./data/aztec20";
