+++ src/kzg10.rs
@@ -1,0 +1,3 @@
+//! SRS related to [KZG10](https://cacr.uwaterloo.ca/techreports/2010/cacr2010-10.pdf)
+
+pub mod aztec20;
